# src/main/webapp/app/shared/api/generated/API-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "OncoKB, a comprehensive and curated precision oncology knowledge base, offers oncologists detailed, evidence-based information about individual somatic mutations and structural alterations present in patient tumors with the goal of supporting optimal treatment decisions.",
    "version": "v1.0.0",
    "title": "OncoKB APIs",
    "contact": {
      "name": "OncoKB",
      "url": "https://www.oncokb.org",
      "email": "contact@oncokb.org"
    },
    "license": {
      "name": "Terms of Use",
      "url": "https://www.oncokb.org/terms"
    }
  },
...
  "paths": {
...
    "/api/account/active-trial/init": {
      "post": {
        "tags": [
          "account-resource"
        ],
        "summary": "initiateTrialAccountActivation",
        "operationId": "initiateTrialAccountActivationUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "mail",
            "description": "mail",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/tokens/{id}": {
      "get": {
        "tags": [
          "token-resource"
        ],
        "summary": "getToken",
        "operationId": "getTokenUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "token-resource"
        ],
        "summary": "deleteToken",
        "operationId": "deleteTokenUsingDELETE_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "user-resource"
        ],
        "summary": "getAllUsers",
        "operationId": "getAllUsersUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number of the requested page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "paged",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "size",
            "in": "query",
            "description": "Size of a page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "sort.sorted",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "sort.unsorted",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "unpaged",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserDTO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "user-resource"
        ],
        "summary": "createUser",
        "operationId": "createUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "managedUserVM",
            "description": "managedUserVM",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ManagedUserVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "user-resource"
        ],
        "summary": "updateUser",
        "operationId": "updateUserUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "sendEmail",
            "in": "query",
            "description": "sendEmail",
            "required": true,
            "type": "boolean"
          },
          {
            "in": "body",
            "name": "userDTO",
            "description": "userDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "Activation": {
      "type": "object",
      "properties": {
        "activationDate": {
          "type": "string",
          "format": "date-time"
        },
        "initiationDate": {
          "type": "string",
          "format": "date-time"
        },
        "key": {
          "type": "string"
        }
      },
      "title": "Activation"
    },
    "AdditionalInfoDTO": {
      "type": "object",
      "properties": {
        "trialAccount": {
          "$ref": "#/definitions/TrialAccount"
        },
        "userCompany": {
          "$ref": "#/definitions/UserCompany"
        }
      },
      "title": "AdditionalInfoDTO"
    },
    "Contact": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      },
      "title": "Contact"
    },
...
    "LicenseAgreement": {
      "type": "object",
      "properties": {
        "acceptanceDate": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      },
      "title": "LicenseAgreement"
    },
...
    "ManagedUserVM": {
      "type": "object",
      "properties": {
        "activated": {
          "type": "boolean"
        },
        "activationKey": {
          "type": "string"
        },
        "additionalInfo": {
          "$ref": "#/definitions/AdditionalInfoDTO"
        },
        "authorities": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "city": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        },
        "email": {
          "type": "string",
          "minLength": 5,
          "maxLength": 254
        },
        "emailVerified": {
          "type": "boolean"
        },
        "firstName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "imageUrl": {
          "type": "string",
          "minLength": 0,
          "maxLength": 256
        },
        "jobTitle": {
          "type": "string"
        },
        "langKey": {
          "type": "string",
          "minLength": 2,
          "maxLength": 10
        },
        "lastModifiedBy": {
          "type": "string"
        },
        "lastModifiedDate": {
          "type": "string",
          "format": "date-time"
        },
        "lastName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "licenseType": {
          "type": "string",
          "enum": [
            "ACADEMIC",
            "COMMERCIAL",
            "RESEARCH_IN_COMMERCIAL",
            "HOSPITAL"
          ]
        },
        "login": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50,
          "pattern": "^[_.@A-Za-z0-9-]*$"
        },
        "notifyUserOnTrialCreation": {
          "type": "boolean"
        },
        "password": {
          "type": "string",
          "minLength": 4,
          "maxLength": 100
        },
        "resetDate": {
          "type": "string",
          "format": "date-time"
        },
        "resetKey": {
          "type": "string"
        },
        "tokenIsRenewable": {
          "type": "boolean"
        },
        "tokenValidDays": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "ManagedUserVM"
    },
...
    "Token": {
      "type": "object",
      "required": [
        "currentUsage",
        "renewable"
      ],
      "properties": {
        "creation": {
          "type": "string",
          "format": "date-time"
        },
        "currentUsage": {
          "type": "integer",
          "format": "int32"
        },
        "expiration": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "renewable": {
          "type": "boolean"
        },
        "token": {
          "type": "string",
          "format": "uuid"
        },
        "usageLimit": {
          "type": "integer",
          "format": "int32"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "title": "Token"
    },
    "TrialAccount": {
      "type": "object",
      "properties": {
        "activation": {
          "$ref": "#/definitions/Activation"
        },
        "licenseAgreement": {
          "$ref": "#/definitions/LicenseAgreement"
        }
      },
      "title": "TrialAccount"
    },
...
    "User": {
      "type": "object",
      "required": [
        "activated",
        "login"
      ],
      "properties": {
        "activated": {
          "type": "boolean"
        },
        "email": {
          "type": "string",
          "minLength": 5,
          "maxLength": 254
        },
        "firstName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "imageUrl": {
          "type": "string",
          "minLength": 0,
          "maxLength": 256
        },
        "langKey": {
          "type": "string",
          "minLength": 2,
          "maxLength": 10
        },
        "lastName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "login": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50,
          "pattern": "^[_.@A-Za-z0-9-]*$"
        },
        "resetDate": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "User"
    },
    "UserCompany": {
      "type": "object",
      "properties": {
        "anticipatedReports": {
          "type": "string"
        },
        "businessContact": {
          "$ref": "#/definitions/Contact"
        },
        "description": {
          "type": "string"
        },
        "size": {
          "type": "string"
        },
        "useCase": {
          "type": "string"
        }
      },
      "title": "UserCompany"
    },
    "UserDTO": {
      "type": "object",
      "properties": {
        "activated": {
          "type": "boolean"
        },
        "activationKey": {
          "type": "string"
        },
        "additionalInfo": {
          "$ref": "#/definitions/AdditionalInfoDTO"
        },
        "authorities": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "city": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        },
        "email": {
          "type": "string",
          "minLength": 5,
          "maxLength": 254
        },
        "emailVerified": {
          "type": "boolean"
        },
        "firstName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "imageUrl": {
          "type": "string",
          "minLength": 0,
          "maxLength": 256
        },
        "jobTitle": {
          "type": "string"
        },
        "langKey": {
          "type": "string",
          "minLength": 2,
          "maxLength": 10
        },
        "lastModifiedBy": {
          "type": "string"
        },
        "lastModifiedDate": {
          "type": "string",
          "format": "date-time"
        },
        "lastName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "licenseType": {
          "type": "string",
          "enum": [
            "ACADEMIC",
            "COMMERCIAL",
            "RESEARCH_IN_COMMERCIAL",
            "HOSPITAL"
          ]
        },
        "login": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50,
          "pattern": "^[_.@A-Za-z0-9-]*$"
        },
        "resetDate": {
          "type": "string",
          "format": "date-time"
        },
        "resetKey": {
          "type": "string"
        }
      },
      "title": "UserDTO"
    },
...
  }
}